@@ -2,11 +2,11 @@ name: google_maps_flutter_web
 description: Web platform implementation of google_maps_flutter
 repository: https://github.com/flutter/packages/tree/main/packages/google_maps_flutter/google_maps_flutter_web
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+maps%22
-version: 0.5.4+3
+version: 0.5.5
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ^3.3.0
+  flutter: ">=3.19.0"
 
 flutter:
   plugin:
@@ -22,10 +22,11 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  google_maps: ^6.1.0
+  google_maps: ^7.1.0
   google_maps_flutter_platform_interface: ^2.4.0
   sanitize_html: ^2.0.0
   stream_transform: ^2.0.0
+  web: ^0.5.0
 
 dev_dependencies:
   flutter_test:
